+++ src/main/scala/org/jesperdj/scalaray/lightsource/DeltaLightSource.scala
@@ -33,9 +33,9 @@ abstract class DeltaLightSource extends LightSource {
 
   // Sample the incident radiance of this light source at the given point (pbrt 14.6.1)
   // Returns the radiance, a ray from the light source to the given point and the value of the probability density for this sample
-  def sampleRadiance(point: Point, u1: Float, u2: Float): (Spectrum, Ray, Float) = { val (rad, ray) = radiance(point); (rad, ray, 1.0f) }
+  def sampleRadiance(point: Point, u1: Double, u2: Double): (Spectrum, Ray, Double) = { val (rad, ray) = radiance(point); (rad, ray, 1.0) }
 
   // Probability density of the direction wi (from the given point to a point on the light source) being sampled with respect to the distribution
-  // that sampleRadiance(point: Point, u1: Float, u2: Float) uses to sample points (pbrt 14.6.1)
-  def pdf(point: Point, wi: Vector): Float = 0.0f
+  // that sampleRadiance(point: Point, u1: Double, u2: Double) uses to sample points (pbrt 14.6.1)
+  def pdf(point: Point, wi: Vector): Double = 0.0
 }
