+++ src/main/scala/org/jesperdj/scalaray/lightsource/SpotLightSource.scala
@@ -23,15 +23,15 @@ import org.jesperdj.scalaray.util._
 import org.jesperdj.scalaray.vecmath._
 
 // Spot light source (pbrt 12.2.1)
-final class SpotLightSource (position: Point, direction: Vector, widthAngle: Float, falloffAngle: Float, intensity: Spectrum) extends DeltaLightSource {
-  require(direction.length > 0.999f && direction.length < 1.001f, "SpotLightSource requires that the direction vector is normalized")
+final class SpotLightSource (position: Point, direction: Vector, widthAngle: Double, falloffAngle: Double, intensity: Spectrum) extends DeltaLightSource {
+  require(direction.length > 0.999 && direction.length < 1.001f, "SpotLightSource requires that the direction vector is normalized")
   require(widthAngle > falloffAngle, "SpotLightSource requires that the width angle must be greater than the falloff angle")
 
-  private val cosWidth = math.cos(widthAngle).toFloat
-  private val cosFalloff = math.cos(falloffAngle).toFloat
+  private val cosWidth = math.cos(widthAngle)
+  private val cosFalloff = math.cos(falloffAngle)
 
   // Create a spot light source using a light-to-world transform
-  def this(lightToWorld: Transform, widthAngle: Float, falloffAngle: Float, intensity: Spectrum) =
+  def this(lightToWorld: Transform, widthAngle: Double, falloffAngle: Double, intensity: Spectrum) =
     this(lightToWorld * Point.Origin, lightToWorld * Vector.ZAxis, widthAngle, falloffAngle, intensity)
 
   // Radiance of this light source at the given point
@@ -41,16 +41,16 @@ final class SpotLightSource (position: Point, direction: Vector, widthAngle: Flo
 
     // Compute falloff factor
     val c = direction * rd.normalize
-    val falloff = if (c < cosWidth) 0.0f else if (c > cosFalloff) 1.0f else {
+    val falloff = if (c < cosWidth) 0.0 else if (c > cosFalloff) 1.0 else {
       val delta = (c - cosWidth) / (cosFalloff - cosWidth)
       delta * delta * delta * delta
     }
 
-    (if (falloff > 0.0f) intensity * (falloff / rd.lengthSquared) else Spectrum.Black, Ray(position, rd, 0.0f, 1.0f))
+    (if (falloff > 0.0) intensity * (falloff / rd.lengthSquared) else Spectrum.Black, Ray(position, rd, 0.0, 1.0))
   }
 
   // Total emitted power of this light source onto the scene
-  def totalPower(scene: Scene): Spectrum = intensity * (2.0f * π * (1.0f - 0.5f * (cosFalloff + cosWidth)))
+  def totalPower(scene: Scene): Spectrum = intensity * (2.0 * π * (1.0 - 0.5 * (cosFalloff + cosWidth)))
 
   override def toString = "SpotLightSource(position=%s, direction=%s, widthAngle=%g, falloffAngle=%g, intensity=%s)" format
     (position, direction, widthAngle, falloffAngle, intensity)
